requirements.txt
biopython==1.79
pandas==1.4.2
streamlit==1.10.0
pillow==9.0.1
bokeh==2.4.1
numpy==1.22.3
seaborn==0.11.2
matplotlib==3.5.1
colorcet==3.0.0
py3dmol==1.8.0
stmol==0.0.7
streamlit-aggrid==0.2.3.post2